pyproject.toml
[tool.poetry]
name = "speechkit"
version = "0.1.0"
description = "Application for make phychologist session protocols"
authors = ["eonae"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
test = "pytest:main"
main = "src.app.entrypoints.main:run_main_async"
worker = "src.app.entrypoints.worker:main"
recognize = "src.cli.cli_recognize:main"
diarize = "src.cli.cli_diarize:main"

[tool.poetry.dependencies]
python = "^3.10"
pyannote-audio = "^3.3.2"
python-dotenv = "^1.1.0"
soundfile = "^0.13.1"
librosa = "^0.11.0"
boto3 = "^1.38.3"
yandex-speechkit = "^1.5.0"
requests = "^2.32.3"
yandex-cloud-ml-sdk = "^0.9.1"
aiogram = "^3.3.0"
celery = "^5.3.0"
redis = "^4.5.0"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0"
pylint = "^3.3.6"
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
bandit = "^1.8.3"
tox = "^4.25.0"
mypy = "^1.8.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/src/app/core/ports/.*\.py$
'''
